@@ -24,27 +24,16 @@ dependencies:
   flutter:
     sdk: flutter
 
-  responsive_framework: ^0.0.15
-
-  # Google Fonts
-  google_fonts: ^1.1.2
-
-  # Date time picker
-  flutter_datetime_picker: ^1.5.0
-
-  # Images 
-  image_picker: ^0.6.7+22
-  path_provider: ^1.6.27
-  shared_preferences: ^0.5.12+4  # Storage for storing simple data on disk
-
-  # Google Maps
-  google_maps_flutter: ^1.2.0
-  geolocator: ^6.2.1  # To get the current location
-  geocoding: ^1.0.5 # To translate the latitude, longtitude to address 
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  responsive_framework: ^0.0.15
+  google_fonts: ^2.1.0
+  flutter_datetime_picker: ^1.5.1
+  image_picker: ^0.8.2
+  google_maps_flutter: ^2.0.6
+  geolocator: ^7.4.0  # To get the current location
+  geocoding: ^2.0.0 # To translate the latitude, longtitude to address 
+  multi_select_flutter: ^4.0.0
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
