@@ -3,22 +3,52 @@ import 'package:haiya_client/shared/models/user_detail.dart';
 class AuthService {
   // TODO: Call API
   Future<UserDetail> signUp(
-    String firstname,
-    String lastname,
-    String gender,
-    String dob,
-    String phone,
-    String address,
     String email,
+    String password,
+    String title,
+    String firstName,
+    String lastName,
+    String gender,
+    String height,
+    String weight,
+    String phone,
+    String addrRoom,
+    String addrFloor,
+    String addrNo,
+    String addrMoo,
+    String addrSoi,
+    String addrRoad,
+    String addrSubDistrict,
+    String addrDistrict,
+    String addrProvince,
+    String addrPostalCode,
+    String dob,
+    List<String> congentialDisease,
+    List<String> drugAllergy,
   ) async {
     currentUser = UserDetail(
-      firstname: firstname,
-      lastname: lastname,
-      gender: gender,
-      dob: dob,
-      phone: phone,
-      address: address,
       email: email,
+      password: password,
+      title: title,
+      firstName: firstName,
+      lastName: lastName,
+      gender: gender,
+      height: height as int,
+      weight: weight as int,
+      phone: phone,
+      addrRoom: addrRoom,
+      addrFloor: addrFloor,
+      addrNo: addrNo,
+      addrMoo: addrMoo,
+      addrSoi: addrSoi,
+      addrRoad: addrRoad,
+      addrSubDistrict: addrSubDistrict,
+      addrDistrict: addrDistrict,
+      addrProvince: addrProvince,
+      addrPostalCode: addrPostalCode,
+      dob: dob,
+      congentialDisease: congentialDisease,
+      drugAllergy: drugAllergy,
     );
     return currentUser;
   }
@@ -28,26 +58,5 @@ class AuthService {
     String password,
   ) async {
     print("Submit signin: " + email + " " + password);
-  }
-
-  Future<UserDetail> updateUserInfo(
-    String firstname,
-    String lastname,
-    String gender,
-    String dob,
-    String phone,
-    String address,
-    String email,
-  ) async {
-    currentUser = UserDetail(
-      firstname: firstname,
-      lastname: lastname,
-      gender: gender,
-      dob: dob,
-      phone: phone,
-      address: address,
-      email: email,
-    );
-    return currentUser;
   }
 }
